+++ tests/Sabre/CalDAV/Schedule/ScheduleDeliverTest.php
@@ -52,7 +52,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="1.2;Message delivered locally":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=1.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -131,7 +131,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="1.2;Message delivered locally":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=1.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -326,7 +326,7 @@ BEGIN:VCALENDAR
 BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
-ORGANIZER;SCHEDULE-STATUS="1.2;Message delivered locally":mailto:user2.sabredav@sabredav.org
+ORGANIZER;SCHEDULE-STATUS=1.2:mailto:user2.sabredav@sabredav.org
 ATTENDEE;PARTSTAT=ACCEPTED:mailto:user2.sabredav@sabredav.org
 ATTENDEE;PARTSTAT=ACCEPTED:mailto:user1.sabredav@sabredav.org
 ATTENDEE:mailto:user3.sabredav@sabredav.org
@@ -364,7 +364,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="3.7;Could not find principal.":mailto:user3.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=3.7:mailto:user3.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -400,7 +400,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="5.2;Could not find local inbox":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=5.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -436,7 +436,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="5.2;Could not locate a calendar-home-set":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=5.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -471,7 +471,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="5.2;Could not find a schedule-default-calendar-URL property":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=5.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -504,7 +504,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="5.2;There was no system capable of delivering the scheduling message":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=5.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;
@@ -539,7 +539,7 @@ BEGIN:VEVENT
 UID:foo
 DTSTART:20140811T230000Z
 ORGANIZER:mailto:user1.sabredav@sabredav.org
-ATTENDEE;SCHEDULE-STATUS="5.2;There was no system capable of delivering the scheduling message":mailto:user2.sabredav@sabredav.org
+ATTENDEE;SCHEDULE-STATUS=5.2:mailto:user2.sabredav@sabredav.org
 END:VEVENT
 END:VCALENDAR
 ICS;